+++ src/vite-env.d.ts
@@ -3,6 +3,8 @@
 interface ImportMetaEnv {
   readonly VITE_APP_NAME: string;
   readonly VITE_APP_VERSION: string;
+  readonly VITE_GITHUB_TOKEN: string;
+  readonly VITE_GITHUB_REPO: string;
 }
 
 interface ImportMeta {